:recycle: refactor(TerminologyMappingGenerator): Separeted DB-related operations from the generator logic
